pyproject.toml
[tool.poetry]
name = "web_bot"
version = "0.1.0"
description = ""
authors = ["Kamran <kamrankek48@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.14.0"
django = "4.2.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
aiohttp = "3.9"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
django-extensions = "^3.2.3"
python-ulid = "^3.0.0"
pillow = "^11.1.0"
d = "^0.2.2"
email-validator = "^2.2.0"
boto3 = "^1.36.24"
django-storages = "^1.14.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
